oblig 2

+++ library-spring-ws-rest-security-oauth/src/main/java/no/hvl/dat152/rest/ws/repository/AuthorRepository.java
@@ -11,7 +11,7 @@ import no.hvl.dat152.rest.ws.model.Author;
 /**
  * @author tdoy
  */
-public interface AuthorRepository extends CrudRepository<Author, Long> {
+public interface AuthorRepository extends CrudRepository<Author, Integer> {
 	
 
 }
